# docs/swagger.yaml
definitions:
...
  request.CreateMaintenanceRequest:
    properties:
      description:
        type: string
      end_at:
        type: string
      monitor_ids:
        items:
          type: integer
        type: array
      name:
        type: string
      start_at:
        type: string
    required:
    - end_at
    - name
    - start_at
    type: object
...
  request.UpdateMaintenanceRequest:
    properties:
      description:
        type: string
      end_at:
        type: string
      monitor_ids:
        items:
          type: integer
        type: array
      name:
        type: string
      start_at:
        type: string
    required:
    - end_at
    - name
    - start_at
    type: object
...
  response.MaintenanceResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      end_at:
        type: string
      id:
        type: integer
      monitor_ids:
        items:
          type: integer
        type: array
      name:
        type: string
      start_at:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  utils.ErrorResponseModel:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
    type: object
  utils.SuccessResponseModel:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/maintenances:
    get:
      description: Get all maintenances. Users see their own, admins see all.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.MaintenanceResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: List maintenances
      tags:
      - Maintenances
    post:
      consumes:
      - application/json
      description: Create a new maintenance window for one or more monitors
      parameters:
      - description: Maintenance details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.CreateMaintenanceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.MaintenanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Create maintenance
      tags:
      - Maintenances
  /api/maintenances/{id}:
    delete:
      description: Delete a maintenance window by ID
      parameters:
      - description: Maintenance ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/utils.SuccessResponseModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Delete maintenance
      tags:
      - Maintenances
    get:
      description: Get a single maintenance by its ID
      parameters:
      - description: Maintenance ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.MaintenanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Get maintenance by ID
      tags:
      - Maintenances
    put:
      consumes:
      - application/json
      description: Update an existing maintenance window by ID
      parameters:
      - description: Maintenance ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated maintenance details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateMaintenanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.MaintenanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Update maintenance
      tags:
      - Maintenances
...
swagger: "2.0"